+++ features/logistic/logisticCommon/src/main/java/com/tokopedia/logisticCommon/data/model/CustomProductLogisticModel.kt
@@ -8,7 +8,7 @@ data class CustomProductLogisticModel (
 data class CPLProductModel(
     var productId: Long = 0,
     var cplStatus: Int = 0,
-    var shipperServices: List<Int> = listOf()
+    var shipperServices: List<Long> = listOf()
 )
 
 data class ShipperListCPLModel(
